Fix the import errors and restructure the components folder
