+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import { Footer } from "@/components/footer"
-import HomePage from "@/components/home-page"
+import { Footer } from "@/components/layout/footer"
+import HomePage from "@/components/pages/home-page"
 
 export default function Home() {
   return (
