+++ src/components/features/algorand/algorand-info.tsx
