+++ src/components/features/chat/guidelines-bubble.tsx
