+++ src/components/layout/header.tsx
@@ -4,10 +4,10 @@ import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { useState, useEffect, useRef } from "react";
 import { Button } from "@/components/ui/button";
-import RuleBuilderModal from "@/components/rule-builder-modal";
+import RuleBuilderModal from "@/components/features/rules/rule-builder-modal";
 import { toast } from "@/hooks/use-toast";
-import { ModeToggle } from "@/components/mode-toggle";
-import AlgorandWalletConnect from "@/components/algorand-wallet-connect";
+import { ModeToggle } from "@/components/shared/mode-toggle";
+import AlgorandWalletConnect from "@/components/features/algorand/algorand-wallet-connect";
 import { useWalletConnection } from "@/components/providers/txnlab-wallet-provider";
 import { cn } from "@/lib/utils";
 import { Menu, X } from "lucide-react";
