+++ src/components/layout/providers.tsx
@@ -2,9 +2,9 @@
 'use client'
 
 import React, { ReactNode } from 'react'
-import { Web3Providers } from './web3/Web3Providers'
-import { ThemeProvider } from './theme-provider'
-import BackgroundPaths from './animated-background'
+import { Web3Providers } from '@/components/web3/Web3Providers'
+import { ThemeProvider } from '@/components/shared/theme-provider'
+import BackgroundPaths from '@/components/shared/animated-background'
 
 export function Providers({ children }: { children: ReactNode }) {
   return (
