+++ src/components/shared/animated-background.tsx
