+++ src/components/shared/background2/floating-paper.tsx
