+++ src/components/shared/background2/sparkles.tsx
