+++ src/components/shared/route-display.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import React, { useState, useMemo } from 'react'
-import { Card, CardContent, CardHeader, CardTitle } from './ui/card'
-import { Button } from './ui/button'
-import { Badge } from './ui/badge'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import { Button } from '@/components/ui/button'
+import { Badge } from '@/components/ui/badge'
 import { formatUnits } from 'viem'
 import { resolveTokenBySymbol } from '@/lib/tokens'
 import { TrendingUp, TrendingDown, ArrowRight, Zap, Clock } from 'lucide-react'
